requirements.txt
pdfplumber==0.11.0
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.3.0
